requirements.txt
fastapi==0.115.6
uvicorn==0.33.0
python-dotenv==1.0.1
python-multipart==0.0.18
langchain==0.3.14
langchain-openai==0.3.0
langchain-pinecone==0.2.0
langchain-community==0.3.14
openai==1.59.3
pinecone-client==5.0.1
pypdf==5.1.0
pytesseract==0.3.13
Pillow==11.1.0
pydantic==2.10.4
